+++ fake-cucumber/javascript/src/SupportCodeExecutor.ts
@@ -3,8 +3,9 @@ import { messages } from '@cucumber/messages'
 import IWorld from './IWorld'
 import { AnyBody } from './types'
 import DataTable from './DataTable'
+import ISupportCodeExecutor from './ISupportCodeExecutor'
 
-export default class SupportCodeExecutor {
+export default class SupportCodeExecutor implements ISupportCodeExecutor {
   constructor(
     public readonly stepDefinitionId: string,
     private readonly body: AnyBody,
@@ -29,20 +30,26 @@ export default class SupportCodeExecutor {
     return this.body.apply(thisObj, argArray)
   }
 
-  public argsToMessages(): messages.IStepMatchArgument[] {
+  public argsToMessages(): messages.TestCase.TestStep.StepMatchArgumentsList.IStepMatchArgument[] {
     return this.args.map((arg) => {
-      return new messages.StepMatchArgument({
-        group: toMessageGroup(arg.group),
-        parameterTypeName: arg.parameterType.name,
-      })
+      return new messages.TestCase.TestStep.StepMatchArgumentsList.StepMatchArgument(
+        {
+          group: toMessageGroup(arg.group),
+          parameterTypeName: arg.parameterType.name,
+        }
+      )
     })
   }
 }
 
-function toMessageGroup(group: Group): messages.StepMatchArgument.IGroup {
-  return new messages.StepMatchArgument.Group({
-    value: group.value,
-    start: group.start,
-    children: group.children.map((g) => toMessageGroup(g)),
-  })
+function toMessageGroup(
+  group: Group
+): messages.TestCase.TestStep.StepMatchArgumentsList.StepMatchArgument.IGroup {
+  return new messages.TestCase.TestStep.StepMatchArgumentsList.StepMatchArgument.Group(
+    {
+      value: group.value,
+      start: group.start,
+      children: group.children.map((g) => toMessageGroup(g)),
+    }
+  )
 }