+++ fake-cucumber/javascript/src/cli.ts
@@ -9,16 +9,13 @@ import {
 } from '@cucumber/gherkin'
 import fs from 'fs'
 import makeFormatStream from './makeFormatStream'
+import { messages } from '@cucumber/messages'
+import makeMeta from './makeMeta'
 
 const program = new Command()
 program.version(packageJson.version)
 program.option('-r, --require <path>', 'override require path')
 program.option('--predictable-ids', 'Use predictable ids', false)
-program.option(
-  '--globals',
-  'Assign Given/When/Then/After/Before to global scope',
-  false
-)
 program.option(
   '-f, --format <format>',
   'output format: ndjson|protobuf',
@@ -27,16 +24,12 @@ program.option(
 
 async function main() {
   program.parse(process.argv)
-  const { globals, predictableIds, format } = program
+  const { predictableIds, format } = program
 
   const paths = program.args
   const requirePaths = program.require ? program.require.split(':') : paths
 
-  const supportCode = await loadSupportCode(
-    predictableIds,
-    requirePaths,
-    globals
-  )
+  const supportCode = await loadSupportCode(predictableIds, requirePaths)
 
   const gherkinOptions: IGherkinOptions = {
     defaultDialect: 'en',
@@ -48,6 +41,9 @@ async function main() {
 
   const envelopeOutputStream = makeFormatStream(format)
   envelopeOutputStream.pipe(process.stdout)
+
+  envelopeOutputStream.write(new messages.Envelope({ meta: makeMeta() }))
+
   const gherkinQuery = new GherkinQuery()
 
   await runCucumber(