+++ gherkin/ruby/lib/gherkin/ast_builder.rb
@@ -146,6 +146,7 @@ module Gherkin
         steps = get_steps(node)
 
         Cucumber::Messages::GherkinDocument::Feature::Background.new(
+          id: @id_generator.new_id,
           location: get_location(background_line, 0),
           keyword: background_line.matched_keyword,
           name: background_line.matched_text,
@@ -180,6 +181,7 @@ module Gherkin
         table_body = rows.nil? ? nil : rows[1..-1]
 
         Cucumber::Messages::GherkinDocument::Feature::Scenario::Examples.new(
+          id: @id_generator.new_id,
           tags: tags,
           location: get_location(examples_line, 0),
           keyword: examples_line.matched_keyword,
@@ -237,6 +239,7 @@ module Gherkin
         description = get_description(header)
 
         Cucumber::Messages::GherkinDocument::Feature::FeatureChild::Rule.new(
+          id: @id_generator.new_id,
           location: get_location(rule_line, 0),
           keyword: rule_line.matched_keyword,
           name: rule_line.matched_text,