+++ json-to-messages/javascript/src/SupportCodeExecutor.ts
@@ -1,0 +1,16 @@
+import { ISupportCodeExecutor } from '@cucumber/fake-cucumber'
+import { messages } from '@cucumber/messages'
+
+class SupportCodeExecutor implements ISupportCodeExecutor {
+  constructor(public readonly stepDefinitionId: string) {}
+
+  execute() {
+    // no-op
+  }
+
+  argsToMessages(): messages.TestCase.TestStep.StepMatchArgumentsList.IStepMatchArgument[] {
+    return []
+  }
+}
+
+export class NilCodeExecutor extends SupportCodeExecutor {}