+++ json-to-messages/javascript/src/index.ts
@@ -1,0 +1,4 @@
+import jsonToMessages from './jsonToMessages'
+import { Implementation } from './types'
+
+export { jsonToMessages, Implementation }