+++ json-to-messages/ruby/features/hooks/hooks_steps.rb
@@ -1,0 +1,19 @@
+Before() do
+  # no-op
+end
+
+Before('@beforeHook') do
+  # no-op
+end
+
+Before('@failBeforeHook') do
+  raise 'Woops !'
+end
+
+After('@afterHook') do
+  # no-op
+end
+
+Given('a passed step') do
+  # no-op
+end