+++ pretty-formatter/go/pretty.go
@@ -7,7 +7,7 @@ package pretty
 
 import (
 	"fmt"
-	"github.com/cucumber/messages-go/v11"
+	"github.com/cucumber/messages-go/v12"
 	gio "github.com/gogo/protobuf/io"
 	"io"
 	"io/ioutil"