+++ query/javascript/test/QueryTest.ts
@@ -3,13 +3,12 @@ import { GherkinStreams, Query as GherkinQuery } from '@cucumber/gherkin'
 import { IdGenerator, messages } from '@cucumber/messages'
 import { pipeline, Readable, Writable } from 'stream'
 import assert from 'assert'
-import SupportCode from '@cucumber/fake-cucumber/dist/src/SupportCode'
-import { withFullStackTrace } from '@cucumber/fake-cucumber/dist/src/ErrorMessageGenerator'
+import { SupportCode, withFullStackTrace } from '@cucumber/fake-cucumber'
 
 import { promisify } from 'util'
 import IncrementClock from '@cucumber/fake-cucumber/dist/src/IncrementClock'
 import Query from '../src/Query'
-import makeTestPlan from '@cucumber/fake-cucumber/dist/src/makeTestPlan'
+import { makeTestPlan, makeTestCase } from '@cucumber/fake-cucumber'
 
 const pipelinePromise = promisify(pipeline)
 
@@ -24,17 +23,20 @@ describe('Query', () => {
   describe('#getWorstTestStepResult(testStepResults)', () => {
     it('returns a FAILED result for PASSED,FAILED,PASSED', () => {
       const result = cucumberQuery.getWorstTestStepResult([
-        new messages.TestStepResult({
-          status: messages.TestStepResult.Status.PASSED,
+        new messages.TestStepFinished.TestStepResult({
+          status: messages.TestStepFinished.TestStepResult.Status.PASSED,
         }),
-        new messages.TestStepResult({
-          status: messages.TestStepResult.Status.FAILED,
+        new messages.TestStepFinished.TestStepResult({
+          status: messages.TestStepFinished.TestStepResult.Status.FAILED,
         }),
-        new messages.TestStepResult({
-          status: messages.TestStepResult.Status.PASSED,
+        new messages.TestStepFinished.TestStepResult({
+          status: messages.TestStepFinished.TestStepResult.Status.PASSED,
         }),
       ])
-      assert.strictEqual(result.status, messages.TestStepResult.Status.FAILED)
+      assert.strictEqual(
+        result.status,
+        messages.TestStepFinished.TestStepResult.Status.FAILED
+      )
     })
   })
 
@@ -43,7 +45,7 @@ describe('Query', () => {
       const results = cucumberQuery.getPickleTestStepResults([])
       assert.deepStrictEqual(
         results.map((r) => r.status),
-        [messages.TestStepResult.Status.UNKNOWN]
+        [messages.TestStepFinished.TestStepResult.Status.UNKNOWN]
       )
     })
 
@@ -65,7 +67,7 @@ describe('Query', () => {
 
       assert.strictEqual(
         testStepResults[0].status,
-        messages.TestStepResult.Status.PASSED
+        messages.TestStepFinished.TestStepResult.Status.PASSED
       )
     })
 
@@ -93,8 +95,8 @@ describe('Query', () => {
       assert.deepStrictEqual(
         testStepResults.map((r) => r.status),
         [
-          messages.TestStepResult.Status.PASSED,
-          messages.TestStepResult.Status.PASSED,
+          messages.TestStepFinished.TestStepResult.Status.PASSED,
+          messages.TestStepFinished.TestStepResult.Status.PASSED,
         ]
       )
     })
@@ -121,7 +123,7 @@ describe('Query', () => {
 
       assert.strictEqual(
         testStepResults[0].status,
-        messages.TestStepResult.Status.UNKNOWN
+        messages.TestStepFinished.TestStepResult.Status.UNKNOWN
       )
     })
   })
@@ -144,8 +146,8 @@ describe('Query', () => {
       assert.deepStrictEqual(
         testStepResults.map((r) => r.status),
         [
-          messages.TestStepResult.Status.PASSED,
-          messages.TestStepResult.Status.FAILED,
+          messages.TestStepFinished.TestStepResult.Status.PASSED,
+          messages.TestStepFinished.TestStepResult.Status.FAILED,
         ]
       )
     })
@@ -169,10 +171,10 @@ describe('Query', () => {
       assert.deepStrictEqual(
         cucumberQuery.getPickleTestStepResults(pickleIds).map((r) => r.status),
         [
-          messages.TestStepResult.Status.PASSED,
-          messages.TestStepResult.Status.PASSED,
-          messages.TestStepResult.Status.PASSED,
-          messages.TestStepResult.Status.FAILED,
+          messages.TestStepFinished.TestStepResult.Status.PASSED,
+          messages.TestStepFinished.TestStepResult.Status.PASSED,
+          messages.TestStepFinished.TestStepResult.Status.PASSED,
+          messages.TestStepFinished.TestStepResult.Status.FAILED,
         ]
       )
     })
@@ -198,8 +200,8 @@ describe('Query', () => {
           )
           .map((r) => r.status),
         [
-          messages.TestStepResult.Status.PASSED,
-          messages.TestStepResult.Status.PASSED,
+          messages.TestStepFinished.TestStepResult.Status.PASSED,
+          messages.TestStepFinished.TestStepResult.Status.PASSED,
         ]
       )
 
@@ -210,8 +212,8 @@ describe('Query', () => {
           )
           .map((r) => r.status),
         [
-          messages.TestStepResult.Status.PASSED,
-          messages.TestStepResult.Status.FAILED,
+          messages.TestStepFinished.TestStepResult.Status.PASSED,
+          messages.TestStepFinished.TestStepResult.Status.FAILED,
         ]
       )
     })
@@ -361,7 +363,7 @@ describe('Query', () => {
         assert.deepEqual(results.length, 1)
         assert.deepEqual(
           results[0].status,
-          messages.TestStepResult.Status.PASSED
+          messages.TestStepFinished.TestStepResult.Status.PASSED
         )
       })
 
@@ -383,7 +385,7 @@ describe('Query', () => {
         assert.deepEqual(results.length, 1)
         assert.deepEqual(
           results[0].status,
-          messages.TestStepResult.Status.PASSED
+          messages.TestStepFinished.TestStepResult.Status.PASSED
         )
       })
     })
@@ -489,8 +491,8 @@ describe('Query', () => {
       queryUpdateStream
     )
 
-    const testPlan = makeTestPlan(gherkinQuery, supportCode)
-    await testPlan.execute((envelope) => {
+    const testPlan = makeTestPlan(gherkinQuery, supportCode, makeTestCase)
+    await testPlan.execute((envelope: messages.IEnvelope) => {
       messagesHandler(envelope)
       cucumberQuery.update(envelope)
     })